+++ csvtk/cmd/filter2.go
@@ -45,6 +45,12 @@ The arithmetic/string expression is supported by:
 
   https://github.com/Knetic/govaluate
 
+Variables formats:
+  $1 or ${1}                        The first field/column
+  $a or ${a}                        Column "a"
+  ${a,b} or ${a b} or ${a (b)}      Column name with special charactors, 
+                                    e.g., commas, spaces, and parentheses
+
 Supported operators and types:
 
   Modifiers: + - / * & | ^ ** % >> <<
@@ -89,11 +95,19 @@ Custom functions:
 		digitsAsString := getFlagBool(cmd, "numeric-as-string")
 
 		fs := make([]string, 0)
+		varType := make(map[string]int)
 		for _, f := range reFilter2.FindAllStringSubmatch(filterStr, -1) {
-			fs = append(fs, f[1])
+			if reFilter2b.MatchString(f[0]) {
+				varType[f[1]] = 1
+				fs = append(fs, f[1])
+			} else {
+				varType[f[2]] = 0
+				fs = append(fs, f[2])
+			}
 		}
 
-		fieldStr := strings.Join(fs, ",")
+		varSep := "__sep__"
+		fieldStr := strings.Join(fs, varSep)
 
 		// ------------------------------------------------------
 
@@ -157,7 +171,7 @@ Custom functions:
 
 		usingColname := true
 
-		fields, colnames, negativeFields, needParseHeaderRow, _ := parseFields(cmd, fieldStr, config.NoHeaderRow)
+		fields, colnames, negativeFields, needParseHeaderRow, _ := parseFields(cmd, fieldStr, varSep, config.NoHeaderRow)
 		var fieldsMap map[int]struct{}
 		if len(fields) > 0 {
 			usingColname = false
@@ -321,7 +335,12 @@ Custom functions:
 					if !usingColname {
 						for _, fieldTmp = range fields {
 							value = record[fieldTmp-1]
-							col = fmt.Sprintf("shenwei%d", fieldTmp)
+							col = strconv.Itoa(fieldTmp)
+							if varType[col] == 1 {
+								col = "${" + col + "}"
+							} else {
+								col = fmt.Sprintf("shenwei%d", fieldTmp)
+							}
 
 							quote = `'`
 
@@ -353,6 +372,8 @@ Custom functions:
 
 							if reFiler2ColSymbolStartsWithDigits.MatchString(col) {
 								col = fmt.Sprintf("shenwei_%s", col)
+							} else if varType[col] == 1 {
+								col = "${" + col + "}"
 							} else {
 								col = "$" + col
 							}
@@ -456,7 +477,8 @@ func init() {
 	filter2Cmd.Flags().BoolP("numeric-as-string", "s", false, `treat even numeric fields as strings to avoid converting big numbers into scientific notation`)
 }
 
-var reFilter2 = regexp.MustCompile(`\$([^ +-/*&\|^%><!~=()"']+)`)
+var reFilter2 = regexp.MustCompile(`\$\{([^}]+?)\}|\$([^ +-/*&\|^%><!~=()"']+)`)
+var reFilter2b = regexp.MustCompile(`^\$\{([^}]+?)\}$`)
 var reFilter2VarField = regexp.MustCompile(`\$(\d+)`)
 
 // var reFilter2VarSymbol = regexp.MustCompile(`\$`)