+++ csvtk/cmd/freq.go
@@ -57,7 +57,7 @@ var freqCmd = &cobra.Command{
 		if fieldStr == "" {
 			checkError(fmt.Errorf("flag -f (--fields) needed"))
 		}
-		fields, colnames, negativeFields, needParseHeaderRow, _ := parseFields(cmd, fieldStr, config.NoHeaderRow)
+		fields, colnames, negativeFields, needParseHeaderRow, _ := parseFields(cmd, fieldStr, ",", config.NoHeaderRow)
 		var fieldsMap map[int]struct{}
 		var fieldsOrder map[int]int      // for set the order of fields
 		var colnamesOrder map[string]int // for set the order of fields