+++ csvtk/cmd/gather.go
@@ -66,7 +66,7 @@ var gatherCmd = &cobra.Command{
 		}
 
 		fuzzyFields := getFlagBool(cmd, "fuzzy-fields")
-		fields, colnames, negativeFields, needParseHeaderRow, _ := parseFields(cmd, fieldStr, config.NoHeaderRow)
+		fields, colnames, negativeFields, needParseHeaderRow, _ := parseFields(cmd, fieldStr, ",", config.NoHeaderRow)
 		var fieldsMap map[int]struct{}
 		var fieldsOrder map[int]int      // for set the order of fields
 		var colnamesOrder map[string]int // for set the order of fields