+++ csvtk/cmd/mutate.go
@@ -70,7 +70,7 @@ var mutateCmd = &cobra.Command{
 		if fieldStr == "" {
 			checkError(fmt.Errorf("flag -f (--fields) needed"))
 		}
-		fields, colnames, negativeFields, needParseHeaderRow, _ := parseFields(cmd, fieldStr, config.NoHeaderRow)
+		fields, colnames, negativeFields, needParseHeaderRow, _ := parseFields(cmd, fieldStr, ",", config.NoHeaderRow)
 		if !(len(fields) == 1 || len(colnames) == 1) {
 			checkError(fmt.Errorf("only single field allowed"))
 		}